Add basic schema to add a license
Write an if else condition and create a flag for each possible license, consisting of an abbreviation for the respective license.
Each flag is associated with an echo command that writes the appropriate license text (still empty) to a LICENSE file.

If the syntax is incorrect, an info message is returned.